+++ lib/shopify_api/admin_versions.rb
@@ -5,6 +5,8 @@ module ShopifyAPI
   module AdminVersions
     SUPPORTED_ADMIN_VERSIONS = T.let([
       "unstable",
+      "2025-07",
+      "2025-04",
       "2025-04",
       "2025-01",
       "2024-10",
@@ -22,8 +24,10 @@ module ShopifyAPI
     ], T::Array[String])
 
     LATEST_SUPPORTED_ADMIN_VERSION = T.let("2025-04", String)
+    RELEASE_CANDIDATE_ADMIN_VERSION = T.let("2025-07", String)
   end
 
   SUPPORTED_ADMIN_VERSIONS = ShopifyAPI::AdminVersions::SUPPORTED_ADMIN_VERSIONS
   LATEST_SUPPORTED_ADMIN_VERSION = ShopifyAPI::AdminVersions::LATEST_SUPPORTED_ADMIN_VERSION
+  RELEASE_CANDIDATE_ADMIN_VERSION = ShopifyAPI::AdminVersions::RELEASE_CANDIDATE_ADMIN_VERSION
 end